Added authentication, REST api
Moved data-management to a separate RESTful app. Added authentication,
though this still needs some work.

App works great now. Still lacks a project portal, but that should
be an easy addition.

+++ cartographer/views.py
@@ -8,18 +8,23 @@ import json
 # Create your views here.
 
 def cartographer(request):
-    if request.method == "GET":
+    if request.user.is_authenticated:
+        if request.method == "GET":
 
-        shapes = DrawnShape.objects.all()
+            shapes = DrawnShape.objects.all()
 
-        context = {"data": serializers.serialize("json",shapes)} 
-        return render(request,"cartographer.html",context)
-    elif request.method == "POST":
-        DrawnShape.objects.all().delete()
-        data = request.body.decode()
-        print(f"Recieved: {data}")
-        shapes = serializers.deserialize("json",data)
-        for shape in shapes: 
-            shape.save()
+            context = {"data": serializers.serialize("json",shapes),
+                    "sessionInfo": {"username":request.user.username,
+                                    "uk":request.user.pk}} 
+            return render(request,"cartographer.html",context)
+        elif request.method == "POST":
+            DrawnShape.objects.all().delete()
+            data = request.body.decode()
+            print(f"Recieved: {data} from {request.user}")
+            shapes = serializers.deserialize("json",data)
+            for shape in shapes: 
+                shape.save()
 
-        return HttpResponse("You posted!!")
+            return HttpResponse("You posted!!")
+    else:
+        return redirect("accounts/login")
